Move Int16Tests
