+++ Tests/Batch1/SimpleTypes/Int16Tests.cs
@@ -1,5 +1,6 @@
 ﻿using Bridge.Test.NUnit;
 using System;
+using System.Globalization;
 
 namespace Bridge.ClientTest.SimpleTypes
 {
@@ -8,19 +9,28 @@ namespace Bridge.ClientTest.SimpleTypes
     public class Int16Tests
     {
         [Test]
-        public void TypePropertiesAreCorrect()
+        public void TypePropertiesAreCorrect_SPI_1717()
         {
             Assert.True((object)(short)0 is short);
             Assert.False((object)0.5 is short);
             Assert.False((object)-32769 is short);
             Assert.False((object)32768 is short);
             Assert.AreEqual("System.Int16", typeof(short).FullName);
-
+            Assert.False(typeof(short).IsClass);
+            Assert.True(typeof(IComparable<short>).IsAssignableFrom(typeof(short)));
+            Assert.True(typeof(IEquatable<short>).IsAssignableFrom(typeof(short)));
+            Assert.True(typeof(IFormattable).IsAssignableFrom(typeof(short)));
             object s = (short)0;
             Assert.True(s is short);
             Assert.True(s is IComparable<short>);
             Assert.True(s is IEquatable<short>);
             Assert.True(s is IFormattable);
+
+            var interfaces = typeof(short).GetInterfaces();
+            Assert.AreEqual(4, interfaces.Length);
+            Assert.True(interfaces.Contains(typeof(IComparable<short>)));
+            Assert.True(interfaces.Contains(typeof(IEquatable<short>)));
+            Assert.True(interfaces.Contains(typeof(IFormattable)));
         }
 
         [Test]
@@ -47,17 +57,29 @@ namespace Bridge.ClientTest.SimpleTypes
 
             checked
             {
-                Assert.Throws(() => { var b = (short)i1; }, err => err is OverflowException);
+                Assert.Throws<OverflowException>(() =>
+                {
+                    var x = (short)i1;
+                }, "-32769 checked");
                 Assert.AreStrictEqual(-32768, (short)i2, "-32768 checked");
                 Assert.AreStrictEqual(5754, (short)i3, "5754 checked");
                 Assert.AreStrictEqual(32767, (short)i4, "32767 checked");
-                Assert.Throws(() => { var b = (short)i5; }, err => err is OverflowException);
+                Assert.Throws<OverflowException>(() =>
+                {
+                    var x = (short)i5;
+                }, "32768 checked");
 
-                Assert.Throws(() => { var b = (short?)ni1; }, err => err is OverflowException);
+                Assert.Throws<OverflowException>(() =>
+                {
+                    var x = (short?)ni1;
+                }, "nullable -32769 checked");
                 Assert.AreStrictEqual(-32768, (short?)ni2, "nullable -32768 checked");
                 Assert.AreStrictEqual(5754, (short?)ni3, "nullable 5754 checked");
                 Assert.AreStrictEqual(32767, (short?)ni4, "nullable 32767 checked");
-                Assert.Throws(() => { var b = (short?)ni5; }, err => err is OverflowException);
+                Assert.Throws<OverflowException>(() =>
+                {
+                    var x = (short?)ni5;
+                }, "nullable 32768 checked");
                 Assert.AreStrictEqual(null, (short?)ni6, "null checked");
             }
         }
@@ -99,13 +121,32 @@ namespace Bridge.ClientTest.SimpleTypes
         }
 
         [Test]
-        public void IFormattableToStringWorks()
+        public void ToStringWithFormatWorks()
         {
             Assert.AreEqual("123", ((short)0x123).ToString("x"));
         }
 
         [Test]
-        public void TryParseWorks()
+        public void ToStringWithFormatAndProviderWorks()
+        {
+            Assert.AreEqual("123", ((short)0x123).ToString("x", CultureInfo.InvariantCulture));
+        }
+
+        [Test]
+        public void IFormattableToStringWorks()
+        {
+            Assert.AreEqual("123", ((IFormattable)((short)0x123)).ToString("x", CultureInfo.InvariantCulture));
+        }
+
+        // Not C# API
+        //[Test]
+        //public void LocaleFormatWorks()
+        //{
+        //    Assert.AreEqual(((short)0x123).LocaleFormat("x"), "123");
+        //}
+
+        [Test]
+        public void TryParseWorks_SPI_1592()
         {
             short numberResult;
             bool result = short.TryParse("234", out numberResult);
@@ -126,11 +167,13 @@ namespace Bridge.ClientTest.SimpleTypes
 
             result = short.TryParse("54768", out numberResult);
             Assert.False(result);
-            Assert.AreEqual(54768, numberResult);
+            // #1592
+            //Assert.AreEqual(0, numberResult);
 
             result = short.TryParse("-55678", out numberResult);
             Assert.False(result);
-            Assert.AreEqual(-55678, numberResult);
+            // #1592
+            //Assert.AreEqual(0, numberResult);
 
             result = short.TryParse("2.5", out numberResult);
             Assert.False(result);
@@ -141,12 +184,12 @@ namespace Bridge.ClientTest.SimpleTypes
         public void ParseWorks()
         {
             Assert.AreEqual(234, short.Parse("234"));
-            Assert.Throws(() => short.Parse(""));
-            Assert.Throws(() => short.Parse(null));
-            Assert.Throws(() => short.Parse("notanumber"));
-            Assert.Throws(() => short.Parse("54768"));
-            Assert.Throws(() => short.Parse("-55678"));
-            Assert.Throws(() => short.Parse("2.5"));
+            Assert.Throws<FormatException>(() => short.Parse(""));
+            Assert.Throws<ArgumentNullException>(() => short.Parse(null));
+            Assert.Throws<FormatException>(() => short.Parse("notanumber"));
+            Assert.Throws<OverflowException>(() => short.Parse("54768"));
+            Assert.Throws<OverflowException>(() => short.Parse("-55678"));
+            Assert.Throws<FormatException>(() => short.Parse("2.5"));
         }
 
         [Test]
